Initial multi-cluster implementation
Using `mcc` the multi-cluster co-ordinator, this change provides global IAM and
global namespace for a collection of minio deployments.

It is required that each minio deployment be configured with the same external
IDP (either LDAP or OpenID based).

The change allows clusters to be linked after being configured with the `mcc`,
to share IAM credentials (i.e. STS and service account credentials), policy
document definitions and user/group policy mappings.

To ensure that policy definition meanings are consistent across separate MinIO
deployments, it is required that bucket names are unique to a cluster; in a
subsequent change, this will be relaxed to allow the same bucket name in
different clusters if a replication configuration exists between them.